pyproject.toml
[tool.poetry]
name = "uno"
version = "0.1.0"
description = ""
authors = ["Mert Armagan Sari <mertarmagan@windowslive.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.1.0"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pre-commit = "^3.5.0"
pylint = "^3.0.2"

[tool.isort]
profile = "black"

[tool.coverage]
    [tool.coverage.run]
    omit = [
        '__init__.py',
        'tests/*',
        '*/tests/*',
    ]

    [tool.coverage.report]
    skip_empty = true
    fail_under = 90

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
